+++ docs/2019-09-15_OfficialDocumentation/class_io_taa_p___serial.js
@@ -1,7 +1,8 @@
 var class_io_taa_p___serial =
 [
     [ "IoTaaP_Serial", "class_io_taa_p___serial.html#a684057b8cb6394df475d4d953117e19d", null ],
-    [ "getLn", "class_io_taa_p___serial.html#ae655e7cbed0bbfe955ba2fce2669e3f1", null ],
+    [ "getString", "class_io_taa_p___serial.html#a0550ddc3815e6aa70938865ec280fda1", null ],
+    [ "getStringUntil", "class_io_taa_p___serial.html#a71fa16df7d456a254e083f4beacda59d", null ],
     [ "init", "class_io_taa_p___serial.html#a22635b52087201897d4ef12b679c1945", null ],
     [ "printLn", "class_io_taa_p___serial.html#a7d2234b89c5ba0b4898ca80275033999", null ]
 ];