+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 All notable changes to this project will be documented in this file.
+
+## [0.0.2] - 2019-09-17
+  
+Feature improvements
+ 
+### Added
+
+getStringUntil() function for reading UART until delimiter received
+ 
+### Changed
+
+Serial function: getLn() to getString() 
+   
+### Fixed
      
 ## [0.0.1] - 2019-09-15